TSE-1265 Branch and Update TSIntegratedExamples

+++ Source/Modeling/Plugins/JoistAreaFeatures/ManipulationFeature.cs
@@ -83,7 +83,7 @@
                 {
                     foreach (var component in Components)
                     {
-                        if (PluginDataHelper.IsBlankValue(_depthOffset.Value))
+                        if (PluginDataHelper.IsBlankValue((double)_depthOffset.Value))
                         {
                             DmCommon.ModifyComponent(component, "DepthOffset", TxModel.NullDoubleValue);
                             _depthOffset.Title = "Depth (Calculated)";
